Add tests for printableStyleTemplateId.
Found a fixed an infinite loop in WebGUI::Style::process

If an asset was requested to be printed who does not have a
printableStyleTemplateId then process would recurse back up the
lineage until it does find an asset with a printableStyleTemplateId.
Eventually, if none is found, it will find the Asset Root, whose
parent is itself, and it will loop forever.

I added code to jump out of the loop if it recurses back to the
assetRoot.



git-svn-id: https://svn.webgui.org/plainblack/WebGUI@3013 ba3075da-5205-0410-b03c-ad9aa4846add