Fixed the failing FileUrl test and clearly document where the FileUrl macro will work and where it will not.
This test failed due to an API change in WebGUI.  Before 7.3, if you created
a Storage object and then created an Asset with that Storage Id, then WebGUI
would do that happily.  This is actually a bug, because you could create multiple
Assets with that one Storage Id, and when _any_ of the Assets was deleted or
rolled back it would delete the Storage location and leave the other Assets
with that Id pointing to an empty Storage object.

So the API change is actually a bug fix, albeit being a side-effect of the
auto-commit work for posts and wikis.



git-svn-id: https://svn.webgui.org/plainblack/WebGUI@3386 ba3075da-5205-0410-b03c-ad9aa4846add