Session->Privilege is not design well to work in embedded applications like the RTE because by default it uses the default user style from settings. This renders a full page, when we really need a message.
I've added an argument to session->privilege->insufficient to use the
empty style and to use a different message with no embedded macros
to process for logging in.  I tested this on the RTE folder adder and
it works fine.  I duplicated this work in the Image add/edit save routine
as well.

Since privilege->insuffient didn't take arguments before, this preserves
the existing API, which is broken.



git-svn-id: https://svn.webgui.org/plainblack/WebGUI@3637 ba3075da-5205-0410-b03c-ad9aa4846add