Fixed a bug in WebGUI::User::profileField, where the check for whether or not a profile field existed in the db always passed because it used quickArray in scalar context.  Since quickArray always returns 1 element, this was always true and it never returned undef like it was supposed to.
To fix this, I added a new method called quickScalar that returns the first column
from the row as a scalar.  It has a basic test in SQL.t.

Then there's the new test in User.t for coverage that exposed this problem, and
I also removed a whole slew of duplicate tests.


git-svn-id: https://svn.webgui.org/plainblack/WebGUI@3842 ba3075da-5205-0410-b03c-ad9aa4846add